Adjust use of splitmix64 when seeding other RNGs (#25)
The function splitmix64_stateless has been altered to provide a method of
returning sequential results from a single seed. This is not what happens
when "seed+1" is used as the original seed.

In addition a splitmix64_r() function is provided which uses a pointer
argument to locate the seed used for sequential calls.